Add files via upload
There are the same sequence of lengths A and B of natural numbers. The minimum value is calculated by subtracting one number from each of A and B, and multiplying the result by two. The goal is to repeat this process for the length of the sequence so that the final cumulative value is minimized.